requirements.txt
Django==2.0.3
django-ckeditor==5.4.0
django-extensions==2.0.7
django-js-asset==1.1.0
django-mptt==0.9.0
django-tinymce==2.7.0
Markdown==2.6.11
Pillow==5.0.0
python-decouple==3.1
python-memcached==1.59
pytz==2018.3
six==1.11.0
